Detect codes with camera.

+++ 3rdparty/qzxing/qzxing/zxing/zxing/common/reedsolomon/ReedSolomonEncoder.h
@@ -4,7 +4,6 @@
 #include <zxing/common/reedsolomon/GenericGFPoly.h>
 #include <zxing/common/reedsolomon/GenericGF.h>
 #include <zxing/common/Array.h>
-#include <zxing/common/Types.h>
 
 namespace zxing {
 
@@ -19,7 +18,7 @@ private:
 public:
     ReedSolomonEncoder(Ref<GenericGF> field);
 
-    void encode(std::vector<byte> &toEncode, int ecBytes);
+    void encode(std::vector<int>& toEncode, int ecBytes);
 };
 
 }
